Rename the WSTPLink struct to `WstpLink` + add type alias for backwards compatibility
The name `WstpLink` follows the standard Rust naming convention, which does not
capitalize acronyms.

The new `WSTPLink` type alias is marked with the #[deprecated(..)] annotation,
so any libraries making use of the old type name will see a warning instructing
them to update to the new name. This will be removed in a subsequent version.

  * Rename the private `WstpLink.link` field to `raw_link`