feature: Pre-generate wstp-sys bindings so that no end-user configuration is necessary
Previously, wstp-sys could use bindgen in it's build.rs to generate Rust
bindings to wstp.h dynamically. This approach has a few drawbacks:

  * It's relatively slow.

    Bindgen is a large library, and compiling
    bindgen and all it's dependencies is a non-trivial compile time
    expense.

  * It's fragile.

    Generating the bindings requires that:
      - A WL product with the WSTP library header source is available
      - That that WL product is easily locatable (in practice, you had
        to always set an environment variable so that the build script
        would know where to look)
      - The clang compiler is available for bindgen to use as a tool
        to parse/analyze the C headers.

This new approach stores pre-generated Rust declarations for the wstp.h library
in the `wstp-sys/generated/` directory. These files will be committed to
source control. The layout of that structure looks like:

	wstp-sys/generated/
		12.3.0/
			MacOSX-x86-64/WSTP_bindings.rs
			.../
		.../

where the folder structure encodes the WL version and platform that the
bindings are tied to. Encoding the version is needed because the wstp.h
library changes slightly over time (almost always in a backwards-compatible
way). Encoding the target platform is necessary because various types (e.g.
C integer sizes) change depending on the platform.

The new docs/Development.md document describes the basic maintenance actions
to be taken to keep these bindings up to date as new WL versions are
released.