cleanup: Deprecated LinkStr::to_str(), replace with LinkStr::as_str()
Roughly speaking, in Rust naming conventions `as_*` functions should
be "cheap" to call, but `to_*` methods imply some kind of non-trivial
calculation is happening.

Using `to_str()` made sense when this method did it's own double-checking
that the string was valid UTF-8 (not a trivial computation), but this
validation no longer happens except by WSTP, so getting LinkStr data
as a `&str` is essentially a trivial cast.