feature: Add wstp::shutdown() function
This is an escape hatch to work around error messages printed after
main() returns, due to background threads started by the WSTP C library
that may still be running at the end of the program.

See added comment for more details and discussion of a better way
to address this in the future.