featfix: Fix a build issue on Apple Silicon
Due to [Cargo #8875][1], it is not currently possible
to link to so-called "fat" libraries. However, in
WL ~v12.3 and later, WSTP is distributed as a fat binary
containing native for for both x86_64 and arm64 targets.

To work around this incompatibility, the macOS `lipo`
command is used to generate a "thin" library, which cargo
is able to successfully link against.

[1]: https://github.com/rust-lang/cargo/issues/8875