hotfix: Fix issues with README.md path changes found during `cargo publish` (#58)
Due to a recent change to move the wstp crate in a subdirectory of this
repository, the path to the README.md file changed.

This meant updating the `readme = ...` path in Cargo.toml.

A more subtle issue was a piece of code in the library that uses include_str!
to test the examples in README.md referenced a path that does not exist
when the crate is published to crates.io. This was fixed by disabling
this logic unless #[cfg(doctest)] is enabled.